apps/ai-service/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# AI/ML
torch>=2.0.0
transformers>=4.47.0
accelerate>=1.2.0
pillow>=10.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.28.0

# Optional: For quantized models
# bitsandbytes>=0.45.0
# auto-gptq>=0.7.0
